+++ src/tests/unit/conf/fn_config/fn_config_point_test.rs
@@ -52,7 +52,7 @@ mod tests {
         };
         let testPoint2 = PointConfig {
             name: format!("/{}/CraneMovement.BoomDown", self_id),
-            _type: PointConfigType::Float,
+            _type: PointConfigType::Double,
             history: PointConfigHistory::Read,
             alarm: None,
             address: None,
